+++ src/applications/ivc-champva/10-7959C/tests/helpers/utilities.unit.spec.js
@@ -3,10 +3,12 @@ import {
   isRequiredFile,
   nameWording,
   nameWordingExt,
-  getObjectsWithAttachmentId,
 } from '../../helpers/utilities';
 import { requiredFiles } from '../../config/constants';
-import { concatStreets } from '../../../shared/utilities';
+import {
+  concatStreets,
+  getObjectsWithAttachmentId,
+} from '../../../shared/utilities';
 
 describe('isRequiredFile', () => {
   it("should return '(Required)' if required file in formContext", () => {
