+++ src/applications/ivc-champva/shared/utilities.js
@@ -1,3 +1,4 @@
+import _ from 'lodash';
 import { waitForShadowRoot } from 'platform/utilities/ui/webComponents';
 
 /**
@@ -155,4 +156,26 @@ export function concatStreets(addr, newLines = false) {
     }
   }
   return updated;
+}
+
+/**
+ * Retrieves an array of objects containing the property 'attachmentId'
+ * from the given object.
+ *
+ * @param {Object} obj - The input object to search for objects with 'attachmentId'.
+ * @returns {Array} - An array containing objects with the 'attachmentId' property.
+ */
+export function getObjectsWithAttachmentId(obj) {
+  const objectsWithAttachmentId = [];
+  _.forEach(obj, value => {
+    if (_.isArray(value)) {
+      _.forEach(value, item => {
+        if (_.isObject(item) && _.has(item, 'attachmentId')) {
+          objectsWithAttachmentId.push(item);
+        }
+      });
+    }
+  });
+
+  return objectsWithAttachmentId;
 }
